requirements.txt
Flask
bcrypt
PyJWT
flask-sqlalchemy
werkzeug
pyotp
flask-cors
qrcode